arreglando build. codigo viejo que no compila por cambios en el tipado fue comentado.
se transforman todos los Dao de interfaces a abstract class porque de esa forma es la unica que Room permite la herencia y puede funcionar la abstraccion de BaseDao, que esta vez va con cuerpo en cada metodo pero simplemente lanzando UnsupportedOperationException, por lo que es una obligacion implementar todos los metodos pero a nivel Runtime, Compile Time

+++ app/src/main/java/cl/duocuc/gymcel/presentacion/AppNavGraph.kt
@@ -44,7 +44,7 @@ fun AppNavGraph(navController: NavHostController, context: Context) {
             "rutina_detalle/{rutinaId}",
             arguments = listOf(navArgument("rutinaId") { type = NavType.IntType })
         ) { navBackStackEntry ->
-            val rutinaId = navBackStackEntry.arguments?.getInt("rutinaId")
+            val rutinaId = navBackStackEntry.arguments?.getLong("rutinaId")
             RutinaDetalleScreen(
                 navController = navController,
                 rutinaId = rutinaId
